requirements.txt
kiteconnect
openai
google-cloud-firestore
google-cloud-secret-manager
pandas
requests
tqdm
python-dotenv

# 🔍 GPT + Embedding
openai
tiktoken
sentence-transformers==2.2.2
huggingface_hub==0.14.1
torch

# 🔎 FAISS Vector Store
faiss-cpu==1.7.4
numpy==1.24.4
regex
proto-plus>=1.22.3
pydantic
six
